+++ spec/swagger_v2/params_array_spec.rb
@@ -43,11 +43,13 @@ describe 'Group Params as Array' do
         "swagger"=>"2.0",
         "produces"=>["application/json"],
         "host"=>"example.org",
+        "tags" => [{"name"=>"groups", "description"=>"Operations about groups"}, {"name"=>"type_given", "description"=>"Operations about type_givens"}],
         "schemes" => ["https", "http"],
         "paths"=>{
           "/groups"=>{
             "post"=>{
               "produces"=>["application/json"],
+              "tags"=>["groups"],
               "responses"=>{"201"=>{"description"=>"created Group"}},
               "parameters"=>[
                 {"in"=>"formData", "name"=>"required_group[][required_param_1]", "description"=>nil, "type"=>"string", "required"=>true, "allowMultiple"=>true},
@@ -68,18 +70,19 @@ describe 'Group Params as Array' do
         "swagger"=>"2.0",
         "produces"=>["application/json"],
         "host"=>"example.org",
-        "schemes" => ["https", "http"],
+        "tags"=>[{"name"=>"groups", "description"=>"Operations about groups"}, {"name"=>"type_given", "description"=>"Operations about type_givens"}],
+        "schemes"=>["https", "http"],
         "paths"=>{
           "/type_given"=>{
             "post"=>{
               "produces"=>["application/json"],
-              "responses"=>{"201"=>{"description"=>"created TypeGiven"}},
               "parameters"=>[
                 {"in"=>"formData", "name"=>"typed_group[][id]", "description"=>"integer given", "type"=>"integer", "required"=>true, "allowMultiple"=>true, "format"=>"int32"},
                 {"in"=>"formData", "name"=>"typed_group[][name]", "description"=>"string given", "type"=>"string", "required"=>true, "allowMultiple"=>true},
                 {"in"=>"formData", "name"=>"typed_group[][email]", "description"=>"email given", "type"=>"string", "required"=>false, "allowMultiple"=>true},
-                {"in"=>"formData", "name"=>"typed_group[][others]", "description"=>nil, "type"=>"integer", "required"=>false, "allowMultiple"=>true, "format"=>"int32", "enum"=>[1, 2, 3]}
-        ]}}}})
+                {"in"=>"formData", "name"=>"typed_group[][others]", "description"=>nil, "type"=>"integer", "required"=>false, "allowMultiple"=>true, "format"=>"int32", "enum"=>[1, 2, 3]}],
+                "tags"=>["type_given"],
+              "responses"=>{"201"=>{"description"=>"created TypeGiven"}}}}}})
     end
   end
 end